examples/market-dashboard/requirements.txt
fastapi
uvicorn[standard]
jinja2
httpx
apscheduler
python-dotenv
python-multipart
pytest
pytest-asyncio
alpaca-py
yfinance
